+++ wiki_rag/search/main.py
@@ -16,7 +16,7 @@ from pathlib import Path
 from dotenv import load_dotenv
 from langchain_core.messages import AIMessageChunk
 from langchain_core.runnables import RunnableConfig
-from langfuse.callback import CallbackHandler
+from langfuse.langchain import CallbackHandler
 
 import wiki_rag.index as index