+++ wiki_rag/server/main.py
@@ -13,7 +13,7 @@ import uvicorn
 
 from dotenv import load_dotenv
 from langchain_core.runnables import RunnableConfig
-from langfuse.callback import CallbackHandler
+from langfuse.langchain import CallbackHandler
 
 import wiki_rag.index as index